+++ Project1/Core/World/WorldAreas/FrontierDef.cs
@@ -19,9 +19,9 @@ namespace Project1.Core.World.WorldAreas
         public int LootWeightRawMaterial = 1;
         public int LootWeightEquipment = 1;
         public int LootWeightCurrency = 1;
-        public readonly int Tier;
+        public readonly byte Tier;
         LootWrapper LootCurrency;
-        public FrontierDef(string name, int tier) : base(name)
+        public FrontierDef(string name, byte tier) : base(name)
         {
             this.Tier = tier;
         }
